graficarsimpson {TrapSon}	R Documentation

graficarsimpson

Description

Genera la gráfica de la funcion integrada por el metodo de Simpson

Usage


graficarSimpson(func, n, a, b)

Arguments

func
Funcion a graficar por medio del metodo de Simpson

n
Numero de particiones para graficar

a
Limite inferior de la integral

b
Limite superior de la integral

Value

Imprime una grafica con las aproximaciones hechas mediante el metodo de Simpson

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }

[Package TrapSon version 0.1.0 Index]