graficartrap {TrapSon}	R Documentation

graficartrap

Description

Genera la grafica de los trapecios correspondientes al calculo de la integral dada

Usage


graficartrap(func, n, a, b)

Arguments

func
Funcion a graficar por medio del metodo del trapecio

n
Numero de particiones para graficar

a
Limite inferior de la integral

b
Limite superior de la integral

Details

Instalar la libreria "ggplot2" para generar los graficos

Value

Genera la grafica de los trapecios correspondientes a la integral

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (func, n, a, b)
{
    if (a == 0) {
        seg <- seq.int(a + 0.001, b, length.out = n)
    }
    else {
        seg <- seq.int(a, b, length.out = n)
    }
    fx <- vector(length = length(seg))
    for (i in 1:length(seg)) {
        fx[i] <- func(seg[i])
    }
    df <- data.frame(xend = seg, y = rep(0, n), yend = fx, yend1 = c(fx[2:n],
        fx[n]), xend1 = c(seg[2:n], seg[n]))
    ggplot(data = df) + stat_function(fun = func, size = 1.05,
        alpha = 0.75, color = "blue") + geom_segment(aes(x = xend,
        y = y, xend = xend, yend = yend)) + geom_ribbon(aes(x = xend,
        ymin = y, ymax = yend), fill = "black", alpha = 0.3) +
        xlim(c(a, b))
  }

[Package TrapSon version 0.1.0 Index]